just rename and pull the issueplatformtimeseriesquerybuilder into it's own file
