+++ src/sentry/search/events/datasets/issue_platform.py
@@ -1,17 +1,17 @@
-from sentry.search.events.datasets.discover import DiscoverDatasetConfig, USER_DISPLAY_ALIAS
-from sentry.search.events.types import SelectType
-from typing import Optional
 from snuba_sdk import Function
 
+from sentry.search.events import builder
+from sentry.search.events.datasets.base import DatasetConfig
+from sentry.search.events.datasets.discover import USER_DISPLAY_ALIAS
+from sentry.search.events.types import SelectType
 
-class IssuePlatformDatasetConfig(DiscoverDatasetConfig):
 
-	def __init__(self, builder):
-		self.builder = builder # I need to make a new builder but that is tuesday colleen's problem
+class IssuePlatformDatasetConfig(DatasetConfig):
+    def __init__(self, builder: builder.QueryBuilder):
+        self.builder = builder
 
-	def _resolve_user_display_alias(self, _: str) -> SelectType:
-		print("meowra meowra")
-		# columns = ["user.email", "user.username", "user.id", "user.ip"]
-		columns = ["user_email", "user_name", "user_id", "ip_address_v4"]
-		return Function("coalesce", [self.builder.column(column) for column in columns], USER_DISPLAY_ALIAS)
-
+    def _resolve_user_display_alias(self, _: str) -> SelectType:
+        columns = ["user_email", "user_name", "user_id", "ip_address_v4"]
+        return Function(
+            "coalesce", [self.builder.column(column) for column in columns], USER_DISPLAY_ALIAS
+        )
